lint: lint some files
